Add new txhub

+++ AElf.ChainController/TxMemPool/IContractTxPool.cs
@@ -5,13 +5,13 @@
         /// <summary>
         /// the maximal number of tx in one block
         /// </summary>
-        ulong Least { get; }
+        int Least { get; }
         
         /// <summary>
         /// the minimal number of tx in one block
         /// </summary>
-        ulong Limit { get; }
+        int Limit { get; }
 
-        void SetBlockVolume(ulong minimal, ulong maximal);
+        void SetBlockVolume(int minimal, int maximal);
     }
 }
